Improve test coverage & introduce highlight ID
